pasa tests

+++ C++/SRC/Legendre.cpp
@@ -10,10 +10,10 @@ void Legendre(int n, int m, double fi, Matrix &pnm, Matrix &dpnm) {
     dpnm(2, 2) = -sqrt(3) * sin(fi);
     // diagonal coefficients
     for (int i = 2; i <= n; i++) {
-        pnm(i + 1, i + 1) = sqrt((2 * i + 1) / (2 * i)) * cos(fi) * pnm(i, i);
+        pnm(i + 1, i + 1) = sqrt(static_cast<double>(2 * i + 1) / (2 * i)) * cos(fi) * pnm(i, i);
     }
     for (int i = 2; i <= n; i++) {
-        dpnm(i + 1, i + 1) = sqrt((2 * i + 1) / (2 * i)) * ((cos(fi) * dpnm(i, i)) -(sin(fi) * pnm(i, i)));
+        dpnm(i + 1, i + 1) = sqrt(static_cast<double>(2 * i + 1) / (2 * i)) * ((cos(fi) * dpnm(i, i)) -(sin(fi) * pnm(i, i)));
     }
     /* horizontal
     first
@@ -33,7 +33,7 @@ void Legendre(int n, int m, double fi, Matrix &pnm, Matrix &dpnm) {
     int k = 2;
     while (1) {
         for (int i = k; i <= n; i++) {
-            pnm(i + 1, j + 1) = sqrt((2 * i + 1) / ((i - j) * (i + j))) *((sqrt(2 * i - 1) * sin(fi) * pnm(i, j + 1))
+            pnm(i + 1, j + 1) = sqrt(static_cast<double>(2 * i + 1) / ((i - j) * (i + j))) *((sqrt(2 * i - 1) * sin(fi) * pnm(i, j + 1))
             -(sqrt(((i + j - 1) * (i - j - 1)) / (2 * i - 3)) * pnm(i - 1, j + 1)));
         }
         j = j + 1;
@@ -46,7 +46,7 @@ void Legendre(int n, int m, double fi, Matrix &pnm, Matrix &dpnm) {
     k = 2;
     while (1) {
         for (int i = k; i <= n; i++) {
-            dpnm(i + 1, j + 1) = sqrt((2 * i + 1) / ((i - j) * (i + j))) *((sqrt(2 * i - 1) * sin(fi) *
+            dpnm(i + 1, j + 1) = sqrt(static_cast<double>(2 * i + 1) / ((i - j) * (i + j))) *((sqrt(2 * i - 1) * sin(fi) *
                                                                             dpnm(i, j + 1))
             +(sqrt(2 * i - 1) * cos(fi) * pnm(i, j + 1)) -
             (sqrt(((i + j - 1) * (i - j - 1)) / (2 * i - 3)) * dpnm(i - 1, j + 1)));
